// src/interface/types.ts
export interface Coverages {
  ao: number | boolean;
  glassProtection: number | boolean;
  bonusProtection: number | boolean;
}

export interface Discounts {
  voucher: number | boolean;
  commercialDiscount: number | boolean;
  adviserDiscount: number | boolean;
  vipDiscount: number | boolean;
  carSurcharge: number | boolean;
}